renamed files and removed a few eslint inline comments
